change FailedList to ResultList
