+++ PicMergeToPdf/MergerParallel.cs
@@ -8,6 +8,7 @@ namespace PicMerge {
 	/// </summary>
 	/// <param name="finish1img">完成一个文件的回调</param>
 	/// <param name="param">参数</param>
+	/// <err frag="0x8002" ack="0003"></err>
 	internal class MergerParallel(Action finish1img, Parameters param) : Merger(param), IMerger {
 
 		/// <summary>
@@ -18,10 +19,6 @@ namespace PicMerge {
 		/// 完成一张图片（其实是一个文件，不论是否是图片）的回调。
 		/// </summary>
 		private readonly Action FinishOneImg = finish1img;
-		/// <summary>
-		/// 无法合入的文件的列表。
-		/// </summary>
-		private readonly List<FailedFile> m_failed = [];
 
 		private class Count(int _v) {
 			public int value = _v;
@@ -56,7 +53,8 @@ namespace PicMerge {
 		/// <param name="files">输入文件的列表</param>
 		/// <param name="title">内定标题</param>
 		/// <returns>无法合入的文件的列表</returns>
-		public virtual List<FailedFile> Process(string outputfilepath, List<string> files, string? title = null) {
+		public virtual List<FileResult> Process(string outputfilepath, List<string> files, string? title = null) {
+			List<FileResult> result = [];
 			m_files.AddRange(files);
 
 			using PdfTarget pdfTarget = new(outputfilepath, title);
@@ -71,7 +69,7 @@ namespace PicMerge {
 			while (true) {
 				bool isEmpty = true;
 				lock (m_loadedImg) {
-					if (m_loadedImg.Count != 0) {
+					if (m_loadedImg.Count > 0) {
 						isEmpty = false;
 						while (m_loadedImg.Count > 0) {
 							imageDatas.Add(m_loadedImg.Dequeue());
@@ -84,17 +82,20 @@ namespace PicMerge {
 				}
 				// Add Image.
 				foreach (var imageData in imageDatas) {
+					string file = files[imgCnt];
+
 					if (imageData == null) {
-						FinishOneImg();
-						imgCnt++;
-						continue;
+						result.Add(new FileResult(0x80020001, file, StrUnsupported));
 					}
-					if (!pdfTarget.AddImage(imageData, ref m_param)) {
-						lock (m_failed) {
-							m_failed.Add(new FailedFile(0x1001, files[imgCnt], "Unable to add into pdf [iText internal problem]."));
-						}
+					else if (!pdfTarget.AddImage(imageData, ref m_param)) {
+						result.Add(new FileResult(0x80020002, file, StrFailedToAdd));
 					}
+					else {
+						result.Add(new FileResult(0x1, file));
+					}
+
 					FinishOneImg();
+
 					imgCnt++;
 				}
 				imageDatas.Clear();
@@ -102,7 +103,7 @@ namespace PicMerge {
 					break;
 			}
 
-			return m_failed;
+			return result;
 		}
 
 		/// <summary>
@@ -126,15 +127,7 @@ namespace PicMerge {
 			string file = m_files[id];
 
 			/// 加载并处理。
-			ImageData? imageData = null;
-			try {
-				imageData = ParaImage(file);
-			}
-			catch (Exception ex) {
-				lock (m_failed) {
-					m_failed.Add(new FailedFile(0x2001, file, $"Failed to load image because: {ex.Message}"));
-				}
-			}
+			ImageData? imageData = ParaImage(file);
 
 			/// Add loaded data into queue.
 			while (true) {
