+++ PicMergeToPdf/MergerSerial.cs
@@ -8,6 +8,7 @@ namespace PicMerge {
 	/// </summary>
 	/// <param name="finish1img">完成一个文件的回调</param>
 	/// <param name="param">参数</param>
+	/// <err frag="0x8001" ack="0006"></err>
 	internal class MergerSerial(Action finish1img, Parameters param) : Merger(param), IMerger {
 
 		/// <summary>
@@ -30,10 +31,9 @@ namespace PicMerge {
 		/// <param name="outputfilepath">输出文件路径</param>
 		/// <param name="files">输入文件的列表</param>
 		/// <param name="title">内定标题</param>
-		/// <returns>无法合入的文件的列表</returns>
-		public virtual List<FailedFile> Process(string outputfilepath, List<string> files, string? title = null) {
-			/// 无法合入的文件的列表。
-			List<FailedFile> failed = [];
+		/// <returns>合入文件之结果之列表</returns>
+		public virtual List<FileResult> Process(string outputfilepath, List<string> files, string? title = null) {
+			List<FileResult> result = [];
 			try {
 				using PdfTarget pdfTarget = new(outputfilepath, title);
 
@@ -43,56 +43,48 @@ namespace PicMerge {
 				for (; i < files.Count; ++i) {
 					string file = files[i];
 
-					try {
-						imageData = LoadImage(file, m_compressTarget);
-					}
-					catch (Exception ex) {
-						failed.Add(new FailedFile(0x2001, file, $"Failed to load image because: {ex.Message}"));
-						FinishOneImg();
-						continue;
-					}
-
+					imageData = LoadImage(file, m_compressTarget);
 					if (imageData != null) {
 						break;
 					}
-					failed.Add(new FailedFile(0x3001, file, "Unsupported type."));
+					result.Add(new FileResult(0x80010001, file, StrUnsupported));
 					FinishOneImg();
 				}
 				if (imageData == null) {
-					return failed;
+					return result;
 				}
+
 				/// 再打开文件开写。这样的话，如果没有可合入的文件，就不会创建出pdf。
-				if (!pdfTarget.AddImage(imageData, ref m_param)) {
-					failed.Add(new FailedFile(0x4001, files[i], "Unable to add into pdf [iText internal problem]."));
+				if (pdfTarget.AddImage(imageData, ref m_param)) {
+					result.Add(new FileResult(0x1, files[i]));
+				}
+				else {
+					result.Add(new FileResult(0x80010002, files[i], StrFailedToAdd));
 				}
 				FinishOneImg();
+
 				for (++i; i < files.Count; ++i) {
 					string file = files[i];
 
-					try {
-						imageData = LoadImage(file, m_compressTarget);
-					}
-					catch (Exception ex) {
-						failed.Add(new FailedFile(0x2001, file, $"Failed to load image because: {ex.Message}"));
-						FinishOneImg();
-						continue;
-					}
+					imageData = LoadImage(file, m_compressTarget);
 
 					if (imageData == null) {
-						failed.Add(new FailedFile(0x3001, file, "Unsupported type."));
-						FinishOneImg();
-						continue;
+						result.Add(new FileResult(0x80010003, file, StrUnsupported));
 					}
-					if (!pdfTarget.AddImage(imageData, ref m_param)) {
-						failed.Add(new FailedFile(0x4001, files[i], "Unable to add into pdf [iText internal problem]."));
+					else if (!pdfTarget.AddImage(imageData, ref m_param)) {
+						result.Add(new FileResult(0x80010004, file, StrFailedToAdd));
 					}
+					else {
+						result.Add(new FileResult(0x1, file));
+					}
+
 					FinishOneImg();
 				}
 			}
 			catch (Exception ex) {
-				failed = [new FailedFile(0x4002, ex.Message, ex.Source ?? "")];
+				result.Add(new FileResult(0x80010005, ex.Source ?? "", ex.Message));
 			}
-			return failed;
+			return result;
 		}
 
 		public void Dispose() {
